Fix custom live Bitcoin example
Since the Bitcoin type derives directly from BaseData and BaseData
uses the same storage space for both the Time and EndTime properties,
the 'coin.Time =' line was constantly rewinding the EndTime far enough
into the past such that the FastForwardEnumerator would attempt to
skip and fast forward to more current data. This caused a never ending
loop of rewinding the time in Bitcoin and fast forwarding the time
from the live trading enumerator stack.